Changed type passed from Fortran to C for MPI_COMM_WORLD from an integer to a c_int to ensure interoperability.

+++ src/ForTrilinos_external_utils.F90
@@ -1,7 +1,7 @@
 module ForTrilinos_external_utils
+#include "ForTrilinos_config.h"
   use iso_c_binding ,only : c_int        ! Kind parameter (precision specifier)
   use ForTrilinos_enums
-#include "ForTrilinos_config.h"
 
   implicit none                          ! Prevent implicit typing
 
@@ -10,13 +10,13 @@ module ForTrilinos_external_utils
 #ifdef HAVE_MPI
 
   ! /*! Create an Epetra_MpiComm from Fortran */
-  ! CT_Epetra_MpiComm_ID_t Epetra_MpiComm_Fortran_Create ( MPI_Fint fcomm );
+  ! CT_Epetra_MpiComm_ID_t Epetra_MpiComm_Fortran_Create ( int fcomm );
 
   type(FT_Epetra_MpiComm_ID_t) function Epetra_MpiComm_Fortran_Create( fcomm ) &
         bind(C,name='Epetra_MpiComm_Fortran_Create')
-    import :: FT_Epetra_MpiComm_ID_t
+    import :: FT_Epetra_MpiComm_ID_t, c_int
 
-    integer,intent(in),value :: fcomm
+    integer(c_int),intent(in),value :: fcomm
   end function
 
 #endif
